move TouchAwareSwitch into new folder "Component"
