+++ app/src/main/res/layout/activity_main.xml
@@ -99,7 +99,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.565">
 
-      <org.bepass.oblivion.TouchAwareSwitch
+      <org.bepass.oblivion.component.TouchAwareSwitch
           android:id="@+id/switch_button"
           android:layout_width="match_parent"
           android:layout_height="match_parent"
